pyproject.toml
[project]
name = "steamlaunch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py>=2.4.0",
]
